requirements.txt
absl-py==0.1.13
astor==0.6.2
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.6.18
botocore==1.9.18
bz2file==0.98
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
DateTime==4.2
docutils==0.14
Flask==0.12.2
Flask-WTF==0.14.2
gast==0.2.0
gensim==3.4.0
google-api-python-client==1.6.6
google-auth==1.4.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.2
konlpy==0.4.4
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.2
oauth2client==4.1.2
oauthlib==2.0.7
peewee==3.2.0
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
scipy==1.0.1
six==1.11.0
smart-open==1.5.7
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
zope.interface==4.4.3
